refactor : chatBot에 member 연관관계 설정 및 @NonNull final 설정
