+++ src/main/java/org/jungppo/bambooforest/chatbot/domain/entity/ChatBotEntity.java
@@ -1,0 +1,59 @@
+package org.jungppo.bambooforest.chatbot.domain.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+import org.jungppo.bambooforest.global.jpa.domain.entity.JpaBaseEntity;
+import org.jungppo.bambooforest.member.domain.entity.MemberEntity;
+
+@Entity
+@Getter
+@Table(name = "chat_bot")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ChatBotEntity extends JpaBaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String url;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    private String description;
+
+    @Column(nullable = false)
+    private String imageUrl;
+
+    @Column(nullable = false)
+    private String prompt;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "member_id", nullable = false)
+    private MemberEntity member;
+
+    @Builder
+    public ChatBotEntity(@NonNull final String url, @NonNull final String name, @NonNull final String description,
+                         @NonNull final String imageUrl, @NonNull final String prompt,
+                         @NonNull final MemberEntity member) {
+        this.url = url;
+        this.name = name;
+        this.description = description;
+        this.imageUrl = imageUrl;
+        this.prompt = prompt;
+        this.member = member;
+    }
+}
